pyproject.toml
[tool.poetry]
name = "twitter-crew-ai"
version = "0.1.0"
description = "A CrewAI-powered Twitter automation framework"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
crewai = "^0.11.0"
tweepy = "^4.14.0"
langchain-community = "^0.0.19"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 